Calculate Anderson et al. Wind

Leverages the stormwindmodel to allow for the approximate calculation wind exposure at individual points.

Usage

calc_anderson_wind(hurdat2, geography, geoid = "GEOID")

Arguments

hurdat2

Output from get_hurdat2()

geography

sf object for the area of interest

geoid

Unique identifier for geography

Value

Maximum wind and gust speed and duration calculated using the stormwindmodel

Details

Centroids

By default this package will take the mean centroid of whatever geography provided to it. If you would like to approximate exposure at the population weighted centroid use centr::mean_center() on your geography before providing it to this function.

Examples

if (FALSE) { # \dontrun{
harvey_anderson_wind <- calc_anderson_wind(
  harvey_hurdat2,
  TXLA_tracts
)
} # }