Download and Filter Outage Data

Downloadeds publically available EAGLE-I data published by Brelsford et al.

Usage

get_eaglei(interval, path = NULL)

Arguments

interval

Dates to get outage data for

path

Path to download files into (if not provided, files will be download into the tempdir())

Value

Maximum precentage and average duration of outage by county

Details

Customer Estimates

Customer estimates are generated by adding the number of households (ACS centered on year of storm) and number of establishments (Census SUSB for year of storm) per county, getting the percentage of households and establishments per county by state, and multiplying the percentage by the total number of customers in a state (EIA year of storm)

Examples

if (FALSE) { # \dontrun{
harvey_eaglei <- get_eaglei(
  interval = interval(harvey_start, harvey_end + days(5)),
  path = "raw-data/eaglei"
)
} # }