Download and Filter Weather Data

Downloads NOAA nClimGrid-Daily data for the specified interval.

Usage

get_nClimGrid(interval, path = NULL)

Arguments

interval

Dates to get weather data for

path

Path to download files into (if not provided, files will be download into the tempdir())

Value

terra raster of climate data over the interval specified

Examples

if (FALSE) { # \dontrun{
harvey_eaglei <- get_eaglei(
  interval = interval(harvey_start, harvey_end + days(5)),
  path = "raw-data/eaglei"
)
} # }